# src/apps/web/public/anss/openapi/saas.openapi.yaml
openapi: "3.1.0"
info:
  title: "OWorker SaaS Starter API"
  version: "0.1.0"
  description: "Generated from src/apps/api/src/capabilities.mjs."
servers:
  - url: "http://localhost:3001"
paths:
...
  /auth/status:
    get:
      operationId: "saas_auth_status_read"
      summary: "Read CLI/API authentication status."
      x-anss-capability-id: "saas.auth.status.read"
      x-anss-safety:
        access: "public-read"
        writes: false
        requiresUserConfirmation: false
      x-anss-input-schema:
        type: "object"
        additionalProperties: false
        required:

        properties:

      x-anss-output-schema:
        type: "object"
        additionalProperties: false
        required:
          - "schema"
          - "service"
          - "authenticated"
          - "mode"
          - "principal"
          - "instructions"
        properties:
          schema:
            type: "string"
            enum:
              - "anss.auth-status/0.1"
          service:
            type: "string"
            description: "Service id."
          authenticated:
            type: "boolean"
            description: "Whether the supplied credential is accepted."
          mode:
            type: "string"
            description: "Authentication check mode."
          principal:
            type:
              - "object"
              - "null"
            additionalProperties: true
            required:
              - "type"
              - "id"
            properties:
              type:
                type: "string"
                description: "Principal type."
              id:
                type: "string"
                description: "Principal id."
            description: "Authenticated principal, when present."
          instructions:
            type: "object"
            additionalProperties: true
            required:

            properties:

      responses:
        200:
          description: "JSON response"
          content:
            application/json:
              schema:
                type: "object"
                additionalProperties: false
                required:
                  - "schema"
                  - "service"
                  - "authenticated"
                  - "mode"
                  - "principal"
                  - "instructions"
                properties:
                  schema:
                    type: "string"
                    enum:
                      - "anss.auth-status/0.1"
                  service:
                    type: "string"
                    description: "Service id."
                  authenticated:
                    type: "boolean"
                    description: "Whether the supplied credential is accepted."
                  mode:
                    type: "string"
                    description: "Authentication check mode."
                  principal:
                    type:
                      - "object"
                      - "null"
                    additionalProperties: true
                    required:
                      - "type"
                      - "id"
                    properties:
                      type:
                        type: "string"
                        description: "Principal type."
                      id:
                        type: "string"
                        description: "Principal id."
                    description: "Authenticated principal, when present."
                  instructions:
                    type: "object"
                    additionalProperties: true
                    required:

                    properties:

        default:
          description: "Error response"
          content:
            application/json:
              schema:
                type: "object"
                additionalProperties: true
                required:
                  - "status"
                  - "detail"
                properties:
                  status:
                    type: "string"
                    enum:
                      - "error"
                  detail:
                    type: "string"
                    description: "Human-readable error detail."
  /account/summary:
    get:
      operationId: "saas_account_summary_read"
      summary: "Read neutral account and subscription summary."
      x-anss-capability-id: "saas.account.summary.read"
      x-anss-safety:
        access: "demo-read"
        writes: false
        requiresUserConfirmation: false
      x-anss-input-schema:
        type: "object"
        additionalProperties: false
        required:

        properties:

      x-anss-output-schema:
        type: "object"
        additionalProperties: false
        required:
          - "account"
          - "subscription"
        properties:
          account:
            type: "object"
            additionalProperties: true
            required:
              - "id"
              - "label"
            properties:
              id:
                type: "string"
                description: "Account id."
              label:
                type: "string"
                description: "Display label."
          subscription:
            type:
              - "object"
              - "null"
            additionalProperties: true
            required:
              - "plan"
              - "status"
            properties:
              plan:
                type: "string"
                description: "Plan id."
              status:
                type: "string"
                description: "Subscription status."
              monthlyAmountCents:
                type: "integer"
                description: "Monthly amount in cents."
              stripeCustomerId:
                type:
                  - "string"
                  - "null"
                description: "Stripe customer id."
              stripeSubscriptionId:
                type:
                  - "string"
                  - "null"
                description: "Stripe subscription id."
              stripePriceId:
                type:
                  - "string"
                  - "null"
                description: "Stripe price id."
              currentPeriodEnd:
                type:
                  - "string"
                  - "null"
                description: "Current period end timestamp."
            description: "Current subscription, when present."
      responses:
        200:
          description: "JSON response"
          content:
            application/json:
              schema:
                type: "object"
                additionalProperties: false
                required:
                  - "account"
                  - "subscription"
                properties:
                  account:
                    type: "object"
                    additionalProperties: true
                    required:
                      - "id"
                      - "label"
                    properties:
                      id:
                        type: "string"
                        description: "Account id."
                      label:
                        type: "string"
                        description: "Display label."
                  subscription:
                    type:
                      - "object"
                      - "null"
                    additionalProperties: true
                    required:
                      - "plan"
                      - "status"
                    properties:
                      plan:
                        type: "string"
                        description: "Plan id."
                      status:
                        type: "string"
                        description: "Subscription status."
                      monthlyAmountCents:
                        type: "integer"
                        description: "Monthly amount in cents."
                      stripeCustomerId:
                        type:
                          - "string"
                          - "null"
                        description: "Stripe customer id."
                      stripeSubscriptionId:
                        type:
                          - "string"
                          - "null"
                        description: "Stripe subscription id."
                      stripePriceId:
                        type:
                          - "string"
                          - "null"
                        description: "Stripe price id."
                      currentPeriodEnd:
                        type:
                          - "string"
                          - "null"
                        description: "Current period end timestamp."
                    description: "Current subscription, when present."
        default:
          description: "Error response"
          content:
            application/json:
              schema:
                type: "object"
                additionalProperties: true
                required:
                  - "status"
                  - "detail"
                properties:
                  status:
                    type: "string"
                    enum:
                      - "error"
                  detail:
                    type: "string"
                    description: "Human-readable error detail."
  /billing/plans:
    get:
      operationId: "saas_billing_plans_read"
      summary: "Read billing plan metadata."
      x-anss-capability-id: "saas.billing.plans.read"
      x-anss-safety:
        access: "public-read"
        writes: false
        requiresUserConfirmation: false
      x-anss-input-schema:
        type: "object"
        additionalProperties: false
        required:

        properties:

      x-anss-output-schema:
        type: "object"
        additionalProperties: false
        required:
          - "plans"
          - "provider"
          - "status"
        properties:
          plans:
            type: "array"
            items:
              type: "object"
              additionalProperties: false
              required:
                - "id"
                - "name"
                - "monthlyAmountCents"
                - "benefits"
              properties:
                id:
                  type: "string"
                  description: "Plan id."
                name:
                  type: "string"
                  description: "Plan name."
                monthlyAmountCents:
                  type: "integer"
                  description: "Monthly amount in cents."
                benefits:
                  type: "array"
                  items:
                    type: "string"
                    description: "Replaceable benefit label."
          provider:
            type: "string"
            description: "Billing provider."
          status:
            type: "string"
            description: "Billing provider configuration status."
      responses:
        200:
          description: "JSON response"
          content:
            application/json:
              schema:
                type: "object"
                additionalProperties: false
                required:
                  - "plans"
                  - "provider"
                  - "status"
                properties:
                  plans:
                    type: "array"
                    items:
                      type: "object"
                      additionalProperties: false
                      required:
                        - "id"
                        - "name"
                        - "monthlyAmountCents"
                        - "benefits"
                      properties:
                        id:
                          type: "string"
                          description: "Plan id."
                        name:
                          type: "string"
                          description: "Plan name."
                        monthlyAmountCents:
                          type: "integer"
                          description: "Monthly amount in cents."
                        benefits:
                          type: "array"
                          items:
                            type: "string"
                            description: "Replaceable benefit label."
                  provider:
                    type: "string"
                    description: "Billing provider."
                  status:
                    type: "string"
                    description: "Billing provider configuration status."
        default:
          description: "Error response"
          content:
            application/json:
              schema:
                type: "object"
                additionalProperties: true
                required:
                  - "status"
                  - "detail"
                properties:
                  status:
                    type: "string"
                    enum:
                      - "error"
                  detail:
                    type: "string"
                    description: "Human-readable error detail."
...
  /anss/capabilities:
    get:
      operationId: "saas_anss_capabilities_read"
      summary: "Read the Agent-readable service capability list."
      x-anss-capability-id: "saas.anss.capabilities.read"
      x-anss-safety:
        access: "public-read"
        writes: false
        requiresUserConfirmation: false
      x-anss-input-schema:
        type: "object"
        additionalProperties: false
        required:

        properties:

      x-anss-output-schema:
        type: "object"
        additionalProperties: false
        required:
          - "schema"
          - "service"
          - "programmableServiceBoundary"
          - "capabilities"
        properties:
          schema:
            type: "string"
            enum:
              - "anss.capabilities/0.1"
          service:
            type: "string"
            description: "Service id."
          programmableServiceBoundary:
            type: "string"
            description: "Programmable service boundary type."
          capabilities:
            type: "array"
            items:
              type: "object"
              additionalProperties: true
              required:

              properties:

      responses:
        200:
          description: "JSON response"
          content:
            application/json:
              schema:
                type: "object"
                additionalProperties: false
                required:
                  - "schema"
                  - "service"
                  - "programmableServiceBoundary"
                  - "capabilities"
                properties:
                  schema:
                    type: "string"
                    enum:
                      - "anss.capabilities/0.1"
                  service:
                    type: "string"
                    description: "Service id."
                  programmableServiceBoundary:
                    type: "string"
                    description: "Programmable service boundary type."
                  capabilities:
                    type: "array"
                    items:
                      type: "object"
                      additionalProperties: true
                      required:

                      properties:

        default:
          description: "Error response"
          content:
            application/json:
              schema:
                type: "object"
                additionalProperties: true
                required:
                  - "status"
                  - "detail"
                properties:
                  status:
                    type: "string"
                    enum:
                      - "error"
                  detail:
                    type: "string"
                    description: "Human-readable error detail."